renamed plugins -> navigators
