+++ nav2_bt_navigator/src/navigators/navigate_through_poses.cpp
@@ -17,7 +17,7 @@
 #include <set>
 #include <memory>
 #include <limits>
-#include "nav2_bt_navigator/plugins/navigate_through_poses.hpp"
+#include "nav2_bt_navigator/navigators/navigate_through_poses.hpp"
 
 namespace nav2_bt_navigator
 {
