+++ nav2_bt_navigator/src/navigators/navigate_to_pose.cpp
@@ -17,7 +17,7 @@
 #include <set>
 #include <memory>
 #include <limits>
-#include "nav2_bt_navigator/plugins/navigate_to_pose.hpp"
+#include "nav2_bt_navigator/navigators/navigate_to_pose.hpp"
 
 namespace nav2_bt_navigator
 {
